+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
 - HTTP: Add `keep_alive: Option<KeepAlive>` and `so_linger: Option<Duration>` to server `Configuration`
 
+## [0.52.1] - 2026-03-10
+
+### Fixed
+- Fix several compilation issues with esp32p4 and esp32h2
+
 ## [0.52.0] - 2026-03-09
 
 ### Breaking